feat(array): add storage backend
