+++ docarray/array/storage/memory/__init__.py
@@ -1,0 +1,7 @@
+from .backend import MemoryBackendMixin
+from .getsetdel import GetSetDelMixin
+from .seqlike import SequenceLikeMixin
+
+
+class MemoryStorageMixins(MemoryBackendMixin, GetSetDelMixin, SequenceLikeMixin):
+    ...
