+++ docarray/array/storage/sqlite/base.py
