+++ docarray/array/storage/sqlite/dict.py
