+++ docarray/array/storage/sqlite/mixin.py
